requirements2.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==2.1.7
django-cors-headers==2.5.2
django-environ==0.4.5
django-extensions==2.1.6
djangorestframework==3.9.2
idna==2.9
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.15.2
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2-binary==2.8
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
traitlets==4.3.3
twilio==6.35.5
urllib3==1.25.8
uWSGI==2.0.18
wcwidth==0.1.8
Werkzeug==0.15.2
